Move components to features, style table[WIP]
- In conversations with @alina-beck, she suggested we move the
ReportsTable and FiledTable to features since it is tightly coupled to
Reports... we can work towards a table that is "dumb" and just takes the
data and displays it in another PR
- Add extra column for Current status
- Fix widths
